Add unit tests for normalizeMultiselectInput in ClickUp, Docusign, and Freshdesk
